+++ test/dialects_ansi_test.py
@@ -57,7 +57,25 @@ def test__dialect__ansi__file_from_raw(raw, res, caplog):
         # Interval literals
         # https://github.com/alanmcruickshank/sqlfluff/issues/148
         ("ExpressionSegment",
-         "DATE_ADD(CURRENT_DATE('America/New_York'), INTERVAL 1 year)")
+         "DATE_ADD(CURRENT_DATE('America/New_York'), INTERVAL 1 year)"),
+        # Array accessors
+        ("ExpressionSegment", "my_array[1]"),
+        ("ExpressionSegment", "my_array[OFFSET(1)]"),
+        ("ExpressionSegment", "my_array[5:8]"),
+        ("ExpressionSegment", "4 + my_array[OFFSET(1)]"),
+        ("ExpressionSegment", "bits[OFFSET(0)] + 7"),
+        ("SelectTargetElementSegment",
+         ("(count_18_24 * bits[OFFSET(0)])"
+          " / audience_size AS relative_abundance")),
+        ("ExpressionSegment",
+         "count_18_24 * bits[OFFSET(0)] + count_25_34"),
+        ("SelectTargetElementSegment",
+         ("(count_18_24 * bits[OFFSET(0)] + count_25_34)"
+          " / audience_size AS relative_abundance")),
+        # Dense math expressions
+        # https://github.com/alanmcruickshank/sqlfluff/issues/178
+        # https://github.com/alanmcruickshank/sqlfluff/issues/179
+        ("SelectStatementSegment", "SELECT t.val/t.id FROM test WHERE id*1.0/id > 0.8"),
     ]
 )
 def test__dialect__ansi_specific_segment_parses(segmentref, raw, caplog):